Move default formatters to another file
